+++ src/hmdinterface/test/OculusTest.cpp
@@ -149,8 +149,8 @@ TEST_F(OculusTest, BasicSquareRenderWindow) {
   hmdDevice->Initialize();
   auto &cfg = hmdDevice->Configuration();
 
-  renderWindow->SetSize(OSSizeMake(cfg.DisplayWidth(), cfg.DisplayHeight()));
-  renderWindow->SetPosition(OSPointMake(cfg.WindowPositionX(), cfg.WindowPositionY()));
+  renderWindow->SetSize(OSSizeMake(static_cast<float>(cfg.DisplayWidth()), static_cast<float>(cfg.DisplayHeight())));
+  renderWindow->SetPosition(OSPointMake(static_cast<float>(cfg.WindowPositionX()), static_cast<float>(cfg.WindowPositionY())));
   
   renderWindow->ProcessEvents();
